Split BitRange into file
